+++ sdk/communication/Azure.Communication.Common/api/Azure.Communication.Common.netstandard2.0.cs
@@ -1,23 +1,23 @@
 namespace Azure.Communication
 {
-    public partial class CallingApplication : Azure.Communication.CommunicationIdentifier
+    public partial class CallingApplicationIdentifier : Azure.Communication.CommunicationIdentifier
     {
-        public CallingApplication(string id) { }
+        public CallingApplicationIdentifier(string id) { }
         public string Id { get { throw null; } }
     }
     public abstract partial class CommunicationIdentifier
     {
         protected CommunicationIdentifier() { }
     }
-    public partial class CommunicationUser : Azure.Communication.CommunicationIdentifier
+    public partial class CommunicationUserIdentifier : Azure.Communication.CommunicationIdentifier
     {
-        public CommunicationUser(string id) { }
+        public CommunicationUserIdentifier(string id) { }
         public string Id { get { throw null; } }
     }
     [System.Diagnostics.DebuggerDisplayAttribute("{Value}")]
-    public partial class PhoneNumber : Azure.Communication.CommunicationIdentifier
+    public partial class PhoneNumberIdentifier : Azure.Communication.CommunicationIdentifier
     {
-        public PhoneNumber(string phoneNumber) { }
+        public PhoneNumberIdentifier(string phoneNumber) { }
         public string Value { get { throw null; } }
     }
     public partial class UnknownIdentifier : Azure.Communication.CommunicationIdentifier