+++ sdk/communication/Azure.Communication.Common/src/CommunicationUserIdentifier.cs
@@ -6,13 +6,13 @@ using Azure.Core;
 namespace Azure.Communication
 {
     /// <summary>Represents a user in Azure Communication Services.</summary>
-    public class CommunicationUser : CommunicationIdentifier
+    public class CommunicationUserIdentifier : CommunicationIdentifier
     {
         /// <summary>The id of the communication user.</summary>
         public string Id { get; }
 
         /// <summary>
-        /// Initializes a new instance of <see cref="CommunicationUser"/>.
+        /// Initializes a new instance of <see cref="CommunicationUserIdentifier"/>.
         /// </summary>
         /// <param name="id">Id of the communication user.</param>
         /// <exception cref="System.ArgumentNullException">
@@ -21,7 +21,7 @@ namespace Azure.Communication
         /// <exception cref="System.ArgumentException">
         /// Thrown when the <paramref name="id"/> is empty.
         /// </exception>
-        public CommunicationUser(string id)
+        public CommunicationUserIdentifier(string id)
         {
             Argument.AssertNotNullOrEmpty(id, nameof(id));
             Id = id;