+++ sdk/communication/Azure.Communication.Common/src/PhoneNumberIdentifier.cs
@@ -8,12 +8,12 @@ namespace Azure.Communication
 {
     /// <summary>Represents a Phone Number.</summary>
     [DebuggerDisplay("{Value}")]
-    public class PhoneNumber : CommunicationIdentifier
+    public class PhoneNumberIdentifier : CommunicationIdentifier
     {
         /// <summary>The phone number in E.164 format.</summary>
         public string Value { get; }
 
-        /// <summary> Initializes a new instance of <see cref="PhoneNumber"/>.</summary>
+        /// <summary> Initializes a new instance of <see cref="PhoneNumberIdentifier"/>.</summary>
         /// <param name="phoneNumber">The phone number in E.164 format.</param>
         /// <exception cref="System.ArgumentNullException">
         /// Thrown when the <paramref name="phoneNumber"/> is null.
@@ -21,7 +21,7 @@ namespace Azure.Communication
         /// <exception cref="System.ArgumentException">
         /// Thrown when the <paramref name="phoneNumber"/> is empty.
         /// </exception>
-        public PhoneNumber(string phoneNumber)
+        public PhoneNumberIdentifier(string phoneNumber)
         {
             Argument.AssertNotNullOrEmpty(phoneNumber, nameof(phoneNumber));
             Value = phoneNumber;